requirements.txt
tqdm
python-dotenv
transformers
anthropic
langchain
langchain-anthropic
